feat(server): replace simple CSRF protection with safe-method CSRF protection
SafeMethodCsrfProtectionHandlerPlugin is a zero-config safeguard for
enabling GET on handlers with SameSite=Lax cookie authentication. It
rejects GET/HEAD requests arriving as cross-site or browser-initiated
top-level navigations, the only contexts where another site can make a
browser attach explicitly marked SameSite=Lax cookies to a safe-method
request, while cookie-less cross-site requests keep working without an
origin allowlist.

SimpleCsrfProtectionHandlerPlugin is removed along with its docs page;
the migration guide now points at the new plugin.
