--- apps/content/docs/plugins/simple-csrf-protection.mdx
@@ -1,97 +1,0 @@
----
-title: "Simple CSRF Protection Plugin"
-description: "Use SimpleCsrfProtectionHandlerPlugin to reject requests that a browser reports as initiated by another site."
-sidebar:
-  label: "Simple CSRF Protection"
----
-
-## How It Works
-
-The plugin decides on [`Sec-Fetch-Site`](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/Sec-Fetch-Site), the [Fetch Metadata](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers#fetch_metadata_request_headers) header reporting who initiated a request:
-
-| `Sec-Fetch-Site` | initiated by | |
-| --- | --- | --- |
-| `same-origin` | your own pages | allowed |
-| `same-site` | another origin on your site | rejected unless listed in `origin` or `allowSameSite` |
-| `cross-site` | another site | rejected unless listed in `origin` |
-| `none` | no page, such as a link opened from an email | rejected |
-
-:::warning[Requests without an `Origin` header]
-Link clicks, address-bar navigations, GET form submissions, and passive loads such as `<img>` send no `Origin`, so no allowlist entry can admit them. A POST form submission does send one and can be allowlisted.
-:::
-
-Requests without Fetch Metadata pass through, such as those from `curl`, mobile apps, and other non-browser clients.
-
-## Limitations
-
-- Fetch Metadata is [Baseline widely available](https://developer.mozilla.org/en-US/docs/Glossary/Baseline/Compatibility), supported by every major browser since Safari 16.4 in March 2023. Older browsers and header-stripping proxies pass through unchecked.
-- Browsers [send Fetch Metadata only to trustworthy URLs](https://w3c.github.io/webappsec-fetch-metadata/): HTTPS and `localhost`. Over plain HTTP the headers are absent while cookies are not, so every request passes, and `localhost` qualifying hides this in development.
-- Keep authentication cookies on `SameSite=Lax` at minimum, and prefer `SameSite=Strict` or a [synchronizer token](https://cheatsheetseries.owasp.org/cheatsheets/Cross-Site_Request_Forgery_Prevention_Cheat_Sheet.html#synchronizer-token-pattern) for high-value requests.
-
-## Setup
-
-```ts twoslash
-import { RPCHandler } from '@orpc/server/fetch'
-import { router } from './shared/planet'
-// ---cut---
-import { SimpleCsrfProtectionHandlerPlugin } from '@orpc/server/plugins'
-
-const handler = new RPCHandler(router, {
-  plugins: [
-    new SimpleCsrfProtectionHandlerPlugin(),
-  ],
-})
-```
-
-:::info
-The `handler` can be any supported oRPC handler, such as [RPCHandler](/docs/rpc/handler), [OpenAPIHandler](/docs/openapi/handler), or a custom one.
-:::
-
-## Trusting Other Origins
-
-Only your own origin is trusted by default. Use the same allowlist you pass to the [CORS Plugin](/docs/plugins/cors):
-
-```ts
-const origin = ['https://app.example.com']
-
-const handler = new RPCHandler(router, {
-  plugins: [
-    new CORSHandlerPlugin({ origin, credentials: true }),
-    new SimpleCsrfProtectionHandlerPlugin({ origin }),
-  ],
-})
-```
-
-`origin` also accepts a single string, or a function. Every entry must be a specific origin, since there is no wildcard:
-
-```ts
-const plugin = new SimpleCsrfProtectionHandlerPlugin({
-  origin: (origin, options) => {
-    return origin.endsWith('.example.com') ? origin : undefined
-  },
-})
-```
-
-Set `allowSameSite: true` to trust your whole site at once, and only when you control every subdomain, since `SameSite` cookies do not distinguish them:
-
-```ts
-const plugin = new SimpleCsrfProtectionHandlerPlugin({
-  allowSameSite: true,
-})
-```
-
-## Restricting Fetch Modes
-
-A trusted site reaches your API by any means: `fetch`, forms, links, and `<img>` all pass. Restrict it to scripted requests if your app renders user-supplied URLs:
-
-```ts
-const plugin = new SimpleCsrfProtectionHandlerPlugin({
-  allowModes: ['cors', 'same-origin'],
-})
-```
-
-`allowModes` matches [`Sec-Fetch-Mode`](https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/Sec-Fetch-Mode): `navigate` for forms and links, `no-cors` for `<img>` and `<script>`. Leave it unset if you rely on HTML forms, including through the [Method Override Plugin](/docs/plugins/method-override). Requests without the header are never rejected.
-
-## Learn More
-
-For implementation details, see the [source code](https://github.com/middleapi/orpc/blob/main/packages/server/src/plugins/simple-csrf-protection.ts).
