+++ apps/content/docs/rpc/handler.mdx
@@ -81,12 +81,12 @@ const handler = new RPCHandler(router, {
 ### Enabling the GET Method
 
 :::danger[Dangerous with cookie-based authentication]
-Enabling `GET` is dangerous when your application stores tokens in cookies with `SameSite=Lax` (the browser default) or `SameSite=None`. These cookies **are still sent on cross-site top-level navigations**, so an attacker only needs a signed-in user to click a link like `https://example.com/rpc/planet/delete?data=...` and the procedure runs with the victim's cookies. No JavaScript, no CORS bypass.
+Enabling `GET` is dangerous when your application stores tokens in cookies not marked `SameSite=Strict`. These cookies **are still sent on cross-site top-level navigations**, so an attacker only needs a signed-in user to click a link like `https://example.com/rpc/planet/delete?data=...` and the procedure runs with the victim's cookies. No JavaScript, no CORS bypass.
 
 To enable `GET` safely, do one of the following:
 
 - Set authentication cookies to `SameSite=Strict`, which browsers never send cross-site
-- Use an independent protection, such as the [Simple CSRF Protection Plugin](/docs/plugins/simple-csrf-protection)
+- Use an independent protection, such as the [Safe Method CSRF Protection Plugin](/docs/plugins/safe-method-csrf-protection)
 - Allow `GET` only for safe procedures that never modify data, as shown below
 
 Learn more about this attack on [MDN](https://developer.mozilla.org/en-US/docs/Web/Security/Attacks/CSRF) and in the [OWASP CSRF Prevention Cheat Sheet](https://cheatsheetseries.owasp.org/cheatsheets/Cross-Site_Request_Forgery_Prevention_Cheat_Sheet.html).
@@ -95,18 +95,18 @@ Learn more about this attack on [MDN](https://developer.mozilla.org/en-US/docs/W
 Add `GET` to `allowMethods` to enable it for every procedure:
 
 ```ts
-import { SimpleCsrfProtectionHandlerPlugin } from '@orpc/server/plugins'
+import { SafeMethodCsrfProtectionHandlerPlugin } from '@orpc/server/plugins'
 import { RPC_DEFAULT_ALLOW_METHODS } from '@orpc/server/standard'
 
 const handler = new RPCHandler(router, {
   allowMethods: ['GET', ...RPC_DEFAULT_ALLOW_METHODS],
   plugins: [
-    new SimpleCsrfProtectionHandlerPlugin(), // reject requests initiated by another site
+    new SafeMethodCsrfProtectionHandlerPlugin(), // reject GET requests that may carry SameSite=Lax cookies
   ],
 })
 ```
 
-The plugin rejects any request the browser reports as initiated elsewhere. List the origins of clients on other sites in its `origin` option.
+The plugin rejects the navigations another site can use to send `SameSite=Lax` cookies, such as link clicks and redirects, while cookie-less requests such as `fetch` keep working. See the [Safe Method CSRF Protection Plugin](/docs/plugins/safe-method-csrf-protection) page for its cookie requirements.
 
 Or pass a function to decide per request. For example, only allow `GET` for procedures that declare it via [OpenAPI metadata](/docs/openapi/routing):
