+++ packages/server/src/plugins/index.test.ts
@@ -4,7 +4,7 @@ it('exports plugins', async () => {
     CORSHandlerPlugin: expect.any(Function),
     RequestHeadersHandlerPlugin: expect.any(Function),
     ResponseHeadersHandlerPlugin: expect.any(Function),
-    SimpleCsrfProtectionHandlerPlugin: expect.any(Function),
+    SafeMethodCsrfProtectionHandlerPlugin: expect.any(Function),
     MethodOverrideHandlerPlugin: expect.any(Function),
     RethrowHandlerPlugin: expect.any(Function),
     RequestCompressionHandlerPlugin: expect.any(Function),
