+++ packages/server/src/plugins/safe-method-csrf-protection.ts
@@ -1,0 +1,72 @@
+import type { StandardHandlerOptions, StandardHandlerPlugin, StandardHandlerRoutingInterceptor, StandardHandlerRoutingInterceptorOptions } from '../adapters/standard'
+import type { Context } from '../context'
+import { toArray } from '@orpc/shared'
+import { flattenStandardHeader } from '@standardserver/core'
+
+/**
+ * Adds Cross-Site Request Forgery (CSRF) protection that makes safe HTTP methods as secure as
+ * unsafe ones such as `POST`. It rejects `GET` and `HEAD` requests arriving as top-level
+ * navigations initiated cross-site or from outside the browser, the only contexts where
+ * another site can make a browser attach `SameSite=Lax` cookies to a safe-method request.
+ *
+ * @remarks
+ * **Note**: Requests browsers send without `SameSite=Lax` cookies, such as cross-site `fetch`
+ * and `<img>`, pass through, so procedures stay reachable from other sites. This safeguard
+ * requires authentication cookies explicitly marked `SameSite=Lax` or `SameSite=Strict`,
+ * since browsers may attach other cookies to the requests that pass.
+ *
+ * @see {@link https://orpc.dev/docs/plugins/safe-method-csrf-protection | Safe Method CSRF Protection Plugin}
+ */
+export class SafeMethodCsrfProtectionHandlerPlugin<T extends Context> implements StandardHandlerPlugin<T> {
+  name = '~safe-method-csrf-protection'
+
+  /** Judge the real request, before batch splits it into client-authored sub-requests. */
+  after = ['~batch']
+
+  init(options: StandardHandlerOptions<T>): StandardHandlerOptions<T> {
+    const routingInterceptor: StandardHandlerRoutingInterceptor<T> = (interceptorOptions) => {
+      if (this.isAllowed(interceptorOptions)) {
+        return interceptorOptions.next()
+      }
+
+      return Promise.resolve({
+        matched: true,
+        response: { status: 403, headers: {}, body: 'Request blocked by CSRF protection.' },
+      })
+    }
+
+    return {
+      ...options,
+      routingInterceptors: [routingInterceptor, ...toArray(options.routingInterceptors)],
+    }
+  }
+
+  private isAllowed({ request }: StandardHandlerRoutingInterceptorOptions<T>): boolean {
+    if (request.method !== 'GET' && request.method !== 'HEAD') {
+      return true
+    }
+
+    const site = flattenStandardHeader(request.headers['sec-fetch-site'])?.toLowerCase()
+
+    // Absent for non-browser clients, browsers too old to send Fetch Metadata, and any browser
+    // over plain HTTP, where the headers are never appended. Nothing to judge in those cases.
+    // Same-site origins are trusted, matching the `SameSite` cookie model where the site is
+    // the trust boundary.
+    if (site === undefined || site === 'same-origin' || site === 'same-site') {
+      return true
+    }
+
+    // Cross-site, and on browser-initiated requests such as email links, `Lax` cookies ride
+    // only top-level navigations: mode `navigate` targeting dest `document`. Embedded
+    // navigations such as `<iframe>` pass, since they are not top-level. Browsers send both
+    // headers together, so a stripped one blocks the request rather than bypassing the check.
+    const mode = flattenStandardHeader(request.headers['sec-fetch-mode'])?.toLowerCase()
+    const dest = flattenStandardHeader(request.headers['sec-fetch-dest'])?.toLowerCase()
+
+    if (mode === undefined || dest === undefined) {
+      return false
+    }
+
+    return mode !== 'navigate' || dest !== 'document'
+  }
+}
